Add custom plugin info modal.
Previously, the plugin info modal used code that was meant for the network
adminstration area.  This meant that regular site administrators could not
click on a plugin link under the "Included Plugins" section to view more
info about a plugin.

Another issue is if the multisite install is using subdomains, the plugin
modal would not render due to the 'X-Frame-Options' header being set to
'sameorigin'.

To workaround these problems, we're introducing our own plugin info modal.
The modal piggybacks off of the install_plugin_information() function.